Better user data db api
